added session repository

+++ src/main/java/com/roj/eztalk/domain/Token.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Token {
-    private Integer token;
+    private Long token;
 }
