+++ src/main/java/com/roj/eztalk/domain/request/SetEmailRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class SetEmailRequest {
-    private Integer token;
+    private Long token;
     private String email;
 }
